feat: crud user

+++ src/main/java/org/example/repository/UsersRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends JpaRepository<UsersEntity, Long> {
+public interface UsersRepository extends JpaRepository<UsersEntity, Integer> {
     UsersEntity getById(int id);
     UsersEntity findByUsername(String username);
 }
